refactor : 약관동의 이동 및 나이 입력 칸에 13세 제한
