+++ .antigravitycli/9e53889b-632c-4ff7-9132-12890a34eee5.json
@@ -1,0 +1,1 @@
+/Users/bugoncha/.gemini/config/projects/9e53889b-632c-4ff7-9132-12890a34eee5.json
