+++ PROJECT_SUMMARY.md
@@ -1,0 +1,86 @@
+# 프로젝트 요약 – rutina_frontend
+
+## 개요
+
+`rutina-frontend`는 개인 루틴을 추적하는 플랫폼을 제공하는 React Native / Expo 애플리케이션입니다. 앱에는 사용자 인증(이메일, 소셜 로그인), AI 기반 루틴 분석, 통계 시각화(히트맵, 차트), 풍부한 시간표/캘린더 UI가 포함됩니다. 코드베이스는 컴포넌트, 훅, 서비스, 스토어, 에셋 등으로 모듈화된 구조를 가지고 있습니다.
+
+## 구현된 주요 기능
+
+- **사용자 관리** – 이메일 회원가입/로그인, 검증, 중복 확인, 카카오·네이버 소셜 로그인. 추가 프로필 필드(나이, 성별, 직업)와 프로필 완성 흐름 추가.
+- **AI 통합** – OpenAI/GPT API와 연동해 루틴 분석, AI 기반 제안, 챗 기반 인사이트 제공. AI 데이터 처리, UI 화면, 서버 동기화 구현.
+- **통계 및 시각화** – 연/월/주 히트맵 API 연동, 통계 탭 및 차트 UI 개선.
+- **시간표·캘린더** – 스와이프 가능한 시간표, 캘린더 선택, 커스텀 카테고리 색상 지원, UI 일관성 확보.
+- **카테고리 관리** – 순서 재배치, 숨김·복구 기능, 서버와 동기화, 커스텀 색상 지원.
+- **네비게이션·레이아웃** – 5개의 탭을 가진 하단 네비게이션 바, 통합 헤더 컴포넌트, 전반적인 UI 폴리시.
+- **로그아웃·세션 관리** – 로그아웃 흐름 및 확인 모달 구현.
+- **프로젝트 설정** – EAS(Expo Application Services) Apple 배포 설정, ESLint 구성, TypeScript 타입 정의, CI 도구 등.
+
+## 아키텍처 하이라이트
+
+- **상태 관리** – `store/`에 커스텀 스토어(Zustand 또는 Redux) 사용.
+- **API 레이어** – `services/`에 위치한 중앙 `api_client.ts`에서 모든 네트워크 요청을 처리하도록 리팩터링.
+- **훅** – 인증, 데이터 조회, AI 상호작용 등을 위한 재사용 가능한 커스텀 훅이 `hooks/`에 존재.
+- **컴포넌트 라이브러리** – UI 컴포넌트는 `components/`에 정리, 헤더·모달·카드 등 여러 화면에서 공유.
+- **스타일링** – 현대적인 그라디언트, 글라스모픽 효과 등을 활용한 Expo 스타일링 적용.
+
+## 설정 및 개발 방법
+
+```bash
+# 의존성 설치
+npm install
+
+# 개발 서버 시작
+npx expo start
+```
+
+- `npm run reset-project` 명령으로 스타터 코드를 `app-example/`으로 이동하고 빈 `app/` 디렉터리를 생성하여 새 프로젝트를 시작할 수 있습니다.
+- 프로젝트는 `eas.json`을 통해 EAS 빌드가 구성되어 있습니다.
+
+## 커밋 히스토리 (선택적 하이라이트)
+
+| 커밋 | 메시지 |
+|--------|---------|
+| 3ce5af1 | Merge pull request #59 – AI 기능 업데이트 |
+| 3c64139 | AI 기반 백엔드 호출 구현 |
+| bcaed23 | 루틴 통계 히트맵 API 연동 |
+| 8a7b167 | 카테고리 순서 변경 및 숨김·복구 서버 동기화 |
+| 5de0cce | EAS를 이용한 Apple 앱 설정 |
+| 61acfde | 리팩터: `api_client.ts` 통합 |
+| abb7011 | 카테고리 및 루틴 API 연동 |
+| 012ab9c | 회원가입 흐름 개선(이메일 인증, 추가 프로필 필드) |
+| 05e8003 | 소셜 로그인 시 추가 유저 정보 입력 단계 생성 |
+| c7ba533 | AI 일정 기능 구현 |
+| 9554596 | UI 일관성 수정 |
+| c3bda20 | AI 데이터 루틴 화면 추가 |
+| 4b875df | GPT‑3.5 API 연동 AI 분석 |
+| 120a99c | 반복 설정 간소화 및 카테고리/스토리지 리팩터링 |
+| 42caf13 | 카카오·네이버 로그인 연동 |
+| 7d2f735 | 회원가입 UX 및 검증 로직 리팩터링 |
+| 6bf61b6 | 이메일 검증 오류 UI 처리 |
+| 4f6ef67 | AI 화면 UI 개선 |
+| 07f529d | 시간표 스와이프 통합·캘린더 분리·통계 UI 개선 |
+| 8da5d9d | 설정 탭 헤더 이동·AI 통계 화면 추가 |
+| 1199b58 | 개인정보 처리 화면, 통합 캘린더 UI |
+| f888a84 | 상세 모달 UI 개선·일정 수정 기능 |
+| 25bd11e | 일정 추가·상세 모달·스와이프 삭제·커스텀 카테고리·색상 구현 |
+| 6814b33 | 로그인 페이지 UI·회원가입 구현 |
+| 79a2f5a | 로그인·회원가입 페이지 생성 및 기능 구현 |
+| f46ac0c | 일정 리스트 날짜 조작·캘린더 라이브러리 교체 |
+| 3eb10dc | 공통 헤더 컴포넌트 추출 |
+| 1eb5db0 | 시간표 UI·현재 요일 자동 선택·현재 시간 빨간선 표시 |
+| f76e7d6 | 5개 탭을 가진 하단 네비게이션 바 |
+| b3ecd7a | Expo 프로젝트 초기 스캐폴드·폰트 설정 |
+| cda6c49 | 최초 커밋 |
+
+*(전체 로그는 `git log` 명령으로 확인 가능)*
+
+## 향후 작업
+
+- 남은 AI 엔드포인트 연결 및 심층 분석 기능 구현.
+- 오프라인 지원 및 데이터 캐시 구현.
+- 단위·통합 테스트 확대.
+- UI 애니메이션 및 접근성 향상.
+
+---
+
+*Generated on 2026‑05‑21.*
